# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/role": {
      "post": {
        "description": "创建自定义角色并可选分配权限。",
        "operationId": "RoleController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "创建角色",
        "tags": [
          "角色管理"
        ]
      },
      "get": {
        "description": "支持关键字、状态、数据范围筛选。",
        "operationId": "RoleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "查询成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "分页查询角色列表",
        "tags": [
          "角色管理"
        ]
      },
      "delete": {
        "description": "禁止删除系统内置角色或仍有成员的角色。",
        "operationId": "RoleController_removeMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "角色 ID 列表",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "批量删除角色",
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/role/permissions": {
      "get": {
        "description": "按模块分组返回全部权限，供角色权限配置使用。",
        "operationId": "RoleController_listPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "查询成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "获取权限列表",
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/role/{id}": {
      "get": {
        "description": "返回角色信息与权限编码列表。",
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "获取角色详情",
        "tags": [
          "角色管理"
        ]
      },
      "patch": {
        "description": "更新角色基础信息与权限。",
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "更新角色",
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/role/{id}/clone": {
      "post": {
        "description": "基于已有角色深拷贝权限与数据边界生成新角色。",
        "operationId": "RoleController_clone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "源角色 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "克隆成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "克隆角色",
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/role/{id}/members": {
      "get": {
        "description": "分页返回已绑定该角色的用户列表。",
        "operationId": "RoleController_listMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "查询角色关联用户",
        "tags": [
          "角色管理"
        ]
      },
      "post": {
        "description": "将指定用户追加绑定到当前角色。",
        "operationId": "RoleController_addMembers",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "用户 ID 列表",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "绑定成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "绑定用户到角色",
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/role/{id}/members/{userId}": {
      "delete": {
        "description": "将指定用户从当前角色解绑。",
        "operationId": "RoleController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "用户 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "解绑成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "解绑角色用户",
        "tags": [
          "角色管理"
        ]
      }
    },
    "/api/role/{id}/permissions": {
      "patch": {
        "description": "覆盖式更新角色的权限列表。",
        "operationId": "RoleController_assignPermissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "角色 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "权限编码列表",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "400": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "401": {
            "description": "未认证或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseSwaggerDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "分配角色权限",
        "tags": [
          "角色管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Admin Zen",
    "description": "Admin Zen API 文档",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "在 Authorization 头中传入 access token，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "ApiErrorDetailSwaggerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true,
            "description": "错误类型名称",
            "example": "UnauthorizedException"
          },
          "message": {
            "type": "string",
            "nullable": true,
            "description": "错误消息",
            "example": "缺少认证令牌"
          },
          "stack": {
            "type": "string",
            "nullable": true,
            "description": "堆栈信息（生产环境通常为空）"
          }
        },
        "required": [
          "name",
          "message",
          "stack"
        ]
      },
      "ApiErrorResponseSwaggerDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "HTTP 状态码",
            "example": 401
          },
          "message": {
            "type": "string",
            "description": "错误摘要",
            "example": "缺少认证令牌"
          },
          "path": {
            "type": "string",
            "description": "请求路径",
            "example": "/api/user"
          },
          "traceId": {
            "type": "string",
            "description": "请求追踪 ID（与响应头 x-trace-id 一致）",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "timestamp": {
            "type": "string",
            "description": "响应时间（ISO 8601）",
            "example": "2026-05-22T08:00:00.000Z"
          },
          "error": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ApiErrorDetailSwaggerDto"
              }
            ]
          },
          "fieldErrors": {
            "type": "object",
            "nullable": true,
            "description": "按字段分组的校验错误",
            "example": {
              "email": [
                "无效的邮箱格式"
              ]
            }
          },
          "formErrors": {
            "nullable": true,
            "description": "顶层表单级校验错误",
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "message",
          "path",
          "traceId",
          "timestamp",
          "error",
          "fieldErrors",
          "formErrors"
        ]
      },
...
    }
  }
}